This elevator conversion system will help tourists understand the Europe and The USA.
As per European standards the floor on the street level is referred to as the first floor which in turn differs from what first floor means in the terms of the American notation. Under the USA convention the first floor is what we understand as the ground/street level floor.
Basic logic to this simple program is that: -
[THE AMERICAN FLOOR NUMBER = THE EUROPEAN FLOOR NUMBER + 1]